Change cephfs test file path logic for fio >=v3.8
Since fio v3.8 the env vars are parsed on the client side instead
of the server side. This means the job options send to all loadgens
are identical. There is no easy way to make a single fio client to
send different job options(such as filename) to different servers.

The solution is to bind mount the different paths on each loadgen
to a universal path, /mnt/benchmaster, before start fio server.